Disable logging on memory tests
This still runs `logM` so we are checking logging memory consumption but we are
not testing memory consumption of the system logging library. This is necessary
because hslogger does not have a predictable memory profile. While logging into
the void or manipulating the loggers memory might grow up to a point. This does
not look like a leak (settles ater a while) but it massively interferes with our
testing because it's footprint is much bigger.